Clean Commit: - complete comments - changed built system
git-svn-id: svn://svn.code.sf.net/p/jreliability/code/trunk@50 0b33cc7e-343a-ee05-a097-9295465598fc

+++ src/de/cs12/reliability/evaluator/SamplingEvaluator.java
@@ -68,7 +68,7 @@ public class SamplingEvaluator<T> extends AbstractEvaluator<T> {
 	 *            the number of samples
 	 * @return the samples
 	 */
-	public Samples getValues(Map<T, Distribution> distributions, double number) {
+	public Samples getValues(Map<T, Distribution> distributions, int number) {
 		IntegralEvaluator<T> evaluator = new IntegralEvaluator<T>(bdd);
 		double high = evaluator.getUpperBound(distributions);
 		double step = high / (double) number;
